refactor: modularize API services and types
