--- src/services/api.ts
@@ -1,524 +1,0 @@
-/** @format */
-
-import {
-  getAccessToken,
-  refreshToken as refreshAuthToken,
-  logout as authLogout,
-} from "./auth";
-import { API_ENDPOINTS } from "@/constants/api";
-import type {
-  ApiResponse,
-  LoginResponse,
-  RefreshResponse,
-  Tenant,
-  User,
-  Role,
-  UserRole,
-  Permission,
-  Plan,
-  Invoice,
-  Payment,
-  Notification,
-} from "@/types/api";
-
-export async function getTenantId(): Promise<string | null> {
-  if (typeof window !== "undefined") {
-    const match = document.cookie.match(/(?:^|; )tenantId=([^;]+)/);
-    return match ? decodeURIComponent(match[1]) : null;
-  }
-  try {
-    const { cookies } = await import("next/headers");
-    return (await cookies()).get("tenantId")?.value ?? null;
-  } catch {
-    return null;
-  }
-}
-
-const BASE_URL = process.env.NEXT_PUBLIC_API_URL;
-const API_PREFIX = "/api";
-
-async function request<T>(
-  path: string,
-  options: RequestInit = {}
-): Promise<ApiResponse<T>> {
-  let accessToken = await getAccessToken();
-
-  const headers = new Headers(options.headers);
-  if (accessToken) {
-    headers.set("Authorization", `Bearer ${accessToken}`);
-  }
-  const tenantId = await getTenantId();
-  if (tenantId) {
-    headers.set("X-Tenant-ID", tenantId);
-  }
-
-  let body = options.body;
-  if (body && typeof body !== "string" && !(body instanceof FormData)) {
-    headers.set("Content-Type", "application/json");
-    body = JSON.stringify(body);
-  }
-
-  try {
-    let res = await fetch(`${BASE_URL}${path}`, { ...options, headers, body });
-
-    if (res.status === 401) {
-      const newToken = await refreshAuthToken();
-      if (newToken) {
-        accessToken = newToken;
-        headers.set("Authorization", `Bearer ${accessToken}`);
-        res = await fetch(`${BASE_URL}${path}`, { ...options, headers, body });
-      } else {
-        await authLogout();
-      }
-    }
-
-    const json = (await res.json().catch(() => null)) as ApiResponse<T> | null;
-
-    if (!res.ok || !json) {
-      return {
-        success: false,
-        message: json?.message || res.statusText || "API request failed",
-        data: null as any,
-        meta: (json?.meta as any) ?? {
-          request_id: "",
-          timestamp: new Date().toISOString(),
-        },
-        errors: json?.errors ?? null,
-      };
-    }
-
-    return json;
-  } catch (err) {
-    console.error("request error:", err);
-    return {
-      success: false,
-      message: err instanceof Error ? err.message : "API request error",
-      data: null as any,
-      meta: {
-        request_id: "",
-        timestamp: new Date().toISOString(),
-      } as any,
-      errors: err,
-    };
-  }
-}
-
-export const api = {
-  get: <T>(path: string, options?: RequestInit) =>
-    request<T>(path, { ...options, method: "GET" }),
-  post: <T>(path: string, body?: any, options?: RequestInit) =>
-    request<T>(path, { ...options, method: "POST", body }),
-  put: <T>(path: string, body?: any, options?: RequestInit) =>
-    request<T>(path, { ...options, method: "PUT", body }),
-  patch: <T>(path: string, body?: any, options?: RequestInit) =>
-    request<T>(path, { ...options, method: "PATCH", body }),
-  delete: <T>(path: string, options?: RequestInit) =>
-    request<T>(path, { ...options, method: "DELETE" }),
-};
-
-export function login(payload: {
-  email: string;
-  password: string;
-}): Promise<ApiResponse<LoginResponse>> {
-  return api.post<LoginResponse>(
-    `${API_PREFIX}${API_ENDPOINTS.auth.login}`,
-    payload,
-  );
-}
-
-export function refreshToken(payload: {
-  refresh_token: string;
-}): Promise<ApiResponse<RefreshResponse>> {
-  return api.post<RefreshResponse>(
-    `${API_PREFIX}${API_ENDPOINTS.auth.refresh}`,
-    payload,
-  );
-}
-
-export function logout(payload: {
-  refresh_token: string;
-}): Promise<ApiResponse<any>> {
-  return api.post<any>(
-    `${API_PREFIX}${API_ENDPOINTS.auth.logout}`,
-    payload,
-  );
-}
-
-export function listTenants(
-  params?: Record<string, string | number>
-): Promise<ApiResponse<Tenant[]>> {
-  const query = params
-    ? `?${new URLSearchParams(params as any).toString()}`
-    : "";
-  return api.get<Tenant[]>(
-    `${API_PREFIX}${API_ENDPOINTS.tenant.list}${query}`,
-  );
-}
-
-export function getTenantByDomain(
-  domain: string
-): Promise<ApiResponse<Tenant>> {
-  return api.get<Tenant>(
-    `${API_PREFIX}${API_ENDPOINTS.tenant.byDomain}?domain=${encodeURIComponent(
-      domain
-    )}`
-  );
-}
-
-export function getTenant(id: string | number): Promise<ApiResponse<Tenant>> {
-  return api.get<Tenant>(
-    `${API_PREFIX}${API_ENDPOINTS.tenant.detail(id)}`,
-  );
-}
-
-export function createTenant(
-  payload: Partial<Tenant>,
-): Promise<ApiResponse<Tenant>> {
-  return api.post<Tenant>(
-    `${API_PREFIX}${API_ENDPOINTS.tenant.list}`,
-    payload,
-  );
-}
-
-export function updateTenant(
-  id: string | number,
-  payload: Partial<Tenant>,
-): Promise<ApiResponse<Tenant>> {
-  return api.patch<Tenant>(
-    `${API_PREFIX}${API_ENDPOINTS.tenant.detail(id)}`,
-    payload,
-  );
-}
-
-export function updateTenantStatus(
-  id: string | number,
-  payload: { status: string },
-): Promise<ApiResponse<Tenant>> {
-  return api.patch<Tenant>(
-    `${API_PREFIX}${API_ENDPOINTS.tenant.status(id)}`,
-    payload,
-  );
-}
-
-export function listTenantUsers(
-  id: string | number,
-  params?: Record<string, string | number>,
-): Promise<ApiResponse<User[]>> {
-  const query = params
-    ? `?${new URLSearchParams(params as any).toString()}`
-    : "";
-  return api.get<User[]>(
-    `${API_PREFIX}${API_ENDPOINTS.tenant.users(id)}${query}`,
-  );
-}
-
-export function addTenantUser(
-  id: string | number,
-  payload: Partial<User> & { role_id?: number; password?: string },
-): Promise<ApiResponse<User>> {
-  return api.post<User>(
-    `${API_PREFIX}${API_ENDPOINTS.tenant.users(id)}`,
-    payload,
-  );
-}
-
-export function listTenantModules(
-  id: string | number,
-  params?: Record<string, string | number>,
-): Promise<ApiResponse<any[]>> {
-  const query = params
-    ? `?${new URLSearchParams(params as any).toString()}`
-    : "";
-  return api.get<any[]>(
-    `${API_PREFIX}${API_ENDPOINTS.tenant.modules(id)}${query}`,
-  );
-}
-
-export function updateTenantModule(
-  id: string | number,
-  payload: { module_id: number; status: string },
-): Promise<ApiResponse<any>> {
-  return api.patch<any>(
-    `${API_PREFIX}${API_ENDPOINTS.tenant.modules(id)}`,
-    payload,
-  );
-}
-
-export function listUsers(
-  params?: Record<string, string | number>
-): Promise<ApiResponse<User[]>> {
-  const query = params
-    ? `?${new URLSearchParams(params as any).toString()}`
-    : "";
-  return api.get<User[]>(
-    `${API_PREFIX}${API_ENDPOINTS.users.list}${query}`,
-  );
-}
-
-export function createUser(payload: Partial<User>): Promise<ApiResponse<User>> {
-  return api.post<User>(
-    `${API_PREFIX}${API_ENDPOINTS.users.list}`,
-    payload,
-  );
-}
-
-export function getUser(
-  id: string | number,
-): Promise<ApiResponse<User>> {
-  return api.get<User>(
-    `${API_PREFIX}${API_ENDPOINTS.users.detail(id)}`,
-  );
-}
-
-export function updateUser(
-  id: string | number,
-  payload: Partial<User>,
-): Promise<ApiResponse<User>> {
-  return api.patch<User>(
-    `${API_PREFIX}${API_ENDPOINTS.users.detail(id)}`,
-    payload,
-  );
-}
-
-export function patchUserStatus(
-  id: string | number,
-  payload: { status: string },
-): Promise<ApiResponse<User>> {
-  return api.patch<User>(
-    `${API_PREFIX}${API_ENDPOINTS.users.status(id)}`,
-    payload,
-  );
-}
-
-export function deleteUser(
-  id: string | number,
-): Promise<ApiResponse<any>> {
-  return api.delete<any>(
-    `${API_PREFIX}${API_ENDPOINTS.users.detail(id)}`,
-  );
-}
-
-export function resetPassword(payload: {
-  email: string;
-  new_password: string;
-}): Promise<ApiResponse<any>> {
-  return api.post<any>(
-    `${API_PREFIX}${API_ENDPOINTS.users.resetPassword}`,
-    payload,
-  );
-}
-
-export function listRoles(): Promise<ApiResponse<Role[]>> {
-  return api.get<Role[]>(`${API_PREFIX}${API_ENDPOINTS.roles.list}`);
-}
-
-export function createRole(
-  payload: Partial<Role>,
-): Promise<ApiResponse<Role>> {
-  return api.post<Role>(
-    `${API_PREFIX}${API_ENDPOINTS.roles.list}`,
-    payload,
-  );
-}
-
-export function updateRole(
-  id: string | number,
-  payload: Partial<Role>,
-): Promise<ApiResponse<Role>> {
-  return api.put<Role>(
-    `${API_PREFIX}${API_ENDPOINTS.roles.detail(id)}`,
-    payload,
-  );
-}
-
-export function deleteRole(
-  id: string | number,
-): Promise<ApiResponse<{ id: number }>> {
-  return api.delete<{ id: number }>(
-    `${API_PREFIX}${API_ENDPOINTS.roles.detail(id)}`,
-  );
-}
-
-export function listRolePermissions(
-  id: string | number,
-): Promise<ApiResponse<Permission[]>> {
-  return api.get<Permission[]>(
-    `${API_PREFIX}${API_ENDPOINTS.roles.permissions(id)}`,
-  );
-}
-
-export function addRolePermission(
-  id: string | number,
-  payload: { obj: string; act: string },
-): Promise<ApiResponse<Permission>> {
-  return api.post<Permission>(
-    `${API_PREFIX}${API_ENDPOINTS.roles.permissions(id)}`,
-    payload,
-  );
-}
-
-export function deleteRolePermission(
-  roleId: string | number,
-  permissionId: string | number,
-): Promise<ApiResponse<{ id: number }>> {
-  return api.delete<{ id: number }>(
-    `${API_PREFIX}${API_ENDPOINTS.roles.permission(roleId, permissionId)}`,
-  );
-}
-
-export function assignRole(
-  userId: string | number,
-  payload: { role_id: string | number; tenant_id?: string | number }
-): Promise<ApiResponse<any>> {
-  return api.post<any>(
-    `${API_PREFIX}${API_ENDPOINTS.users.roles(userId)}`,
-    payload,
-  );
-}
-
-export function listUserRoles(
-  userId: string | number,
-): Promise<ApiResponse<UserRole[]>> {
-  return api.get<UserRole[]>(
-    `${API_PREFIX}${API_ENDPOINTS.users.roles(userId)}`,
-  );
-}
-
-export function removeUserRole(
-  userId: string | number,
-  roleId: string | number,
-): Promise<ApiResponse<{ id: number }>> {
-  return api.delete<{ id: number }>(
-    `${API_PREFIX}${API_ENDPOINTS.users.role(userId, roleId)}`,
-  );
-}
-
-export function listNotifications(
-  params?: Record<string, string | number>
-): Promise<ApiResponse<Notification[]>> {
-  const query = params
-    ? `?${new URLSearchParams(params as any).toString()}`
-    : "";
-  return api.get<Notification[]>(
-    `${API_PREFIX}${API_ENDPOINTS.notifications.list}${query}`,
-  );
-}
-
-export function createNotification(
-  payload: Partial<Notification>
-): Promise<ApiResponse<Notification>> {
-  return api.post<Notification>(
-    `${API_PREFIX}${API_ENDPOINTS.notifications.create}`,
-    payload,
-  );
-}
-
-export function updateNotificationStatus(
-  id: string | number,
-  payload: { status: string }
-): Promise<ApiResponse<Notification>> {
-  return api.patch<Notification>(
-    `${API_PREFIX}${API_ENDPOINTS.notifications.status(id)}`,
-    payload,
-  );
-}
-
-export function listVendorPlans(): Promise<ApiResponse<Plan[]>> {
-  return api.get<Plan[]>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.vendor.plans}`,
-  );
-}
-
-export function createVendorPlan(
-  payload: Partial<Plan>,
-): Promise<ApiResponse<Plan>> {
-  return api.post<Plan>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.vendor.plans}`,
-    payload,
-  );
-}
-
-export function getVendorPlan(
-  id: string | number,
-): Promise<ApiResponse<Plan>> {
-  return api.get<Plan>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.vendor.plan(id)}`,
-  );
-}
-
-export function updateVendorPlan(
-  id: string | number,
-  payload: Partial<Plan>,
-): Promise<ApiResponse<Plan>> {
-  return api.put<Plan>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.vendor.plan(id)}`,
-    payload,
-  );
-}
-
-export function deleteVendorPlan(
-  id: string | number,
-): Promise<ApiResponse<any>> {
-  return api.delete<any>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.vendor.plan(id)}`,
-  );
-}
-
-export function listVendorInvoices(): Promise<ApiResponse<Invoice[]>> {
-  return api.get<Invoice[]>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.vendor.invoices}`,
-  );
-}
-
-export function createVendorInvoice(
-  payload: Partial<Invoice>,
-): Promise<ApiResponse<Invoice>> {
-  return api.post<Invoice>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.vendor.invoices}`,
-    payload,
-  );
-}
-
-export function updateVendorInvoice(
-  id: string | number,
-  payload: Partial<Invoice>,
-): Promise<ApiResponse<Invoice>> {
-  return api.put<Invoice>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.vendor.invoice(id)}`,
-    payload,
-  );
-}
-
-export function deleteVendorInvoice(
-  id: string | number,
-): Promise<ApiResponse<any>> {
-  return api.delete<any>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.vendor.invoice(id)}`,
-  );
-}
-
-export function listClientInvoices(): Promise<ApiResponse<Invoice[]>> {
-  return api.get<Invoice[]>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.client.invoices}`,
-  );
-}
-
-export function createPayment(
-  invoiceId: string | number,
-  payload: Partial<Payment>
-): Promise<ApiResponse<Payment>> {
-  return api.post<Payment>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.client.invoice(invoiceId).payments}`,
-    payload
-  );
-}
-
-export function verifyVendorPayment(
-  id: string | number,
-): Promise<ApiResponse<Payment>> {
-  return api.patch<Payment>(
-    `${API_PREFIX}${API_ENDPOINTS.billing.vendor.payments(id).verify}`,
-  );
-}
-
-export default api;
