+++ src/services/api/index.ts
@@ -1,0 +1,11 @@
+/** @format */
+
+export { api, getTenantId, API_PREFIX } from "./base";
+export { default } from "./base";
+
+export * from "./auth";
+export * from "./tenants";
+export * from "./users";
+export * from "./roles";
+export * from "./notifications";
+export * from "./billing";
