+++ src/types/api/index.ts
@@ -1,0 +1,9 @@
+/** @format */
+
+export * from "./common";
+export * from "./auth";
+export * from "./billing";
+export * from "./tenant";
+export * from "./role";
+export * from "./user";
+export * from "./notification";
