+++ src/types/api/notification.ts
@@ -1,0 +1,10 @@
+/** @format */
+
+export interface Notification {
+  id: number;
+  title: string;
+  message: string;
+  status: string;
+  created_at: string;
+  updated_at: string;
+}
